chore: fixed some global variables
