+++ src/handle/exec.rs
@@ -7,11 +7,12 @@ use hyprland::prelude::HyprDataActive;
 use hyprland::shared::{Address, MonitorId, WorkspaceId};
 use tracing::{debug, span, warn, Level};
 
-use crate::{global, to_client_address, Active, FindByFirst, HyprlandData, Warn};
+use crate::{to_client_address, Active, FindByFirst, HyprlandData};
 
 pub fn switch_to_active(
     active: Option<&Active>,
     clients_data: &HyprlandData,
+    dry: bool,
 ) -> anyhow::Result<()> {
     let _span = span!(Level::TRACE, "exec", active = ?active).entered();
     match active {
@@ -30,24 +31,13 @@ pub fn switch_to_active(
                     id: data.workspace,
                     name: workspace_data.name.clone(),
                 },
-                global::OPTS
-                    .get()
-                    .map(|o| o.dry)
-                    .warn("Failed to access global dry")
-                    .unwrap_or(false),
+                dry,
             )
             .with_context(|| {
                 format!("Failed to execute switch workspace with workspace_data {workspace_data:?}")
             })?;
-            switch_client(
-                &to_client_address(*addr),
-                global::OPTS
-                    .get()
-                    .map(|o| o.dry)
-                    .warn("Failed to access global dry")
-                    .unwrap_or(false),
-            )
-            .with_context(|| format!("Failed to execute with addr {addr:?}"))?;
+            switch_client(&to_client_address(*addr), dry)
+                .with_context(|| format!("Failed to execute with addr {addr:?}"))?;
         }
         Some(Active::Workspace(wid)) => {
             let workspace_data = clients_data
@@ -59,26 +49,16 @@ pub fn switch_to_active(
                     id: *wid,
                     name: workspace_data.name.clone(),
                 },
-                global::OPTS
-                    .get()
-                    .map(|o| o.dry)
-                    .warn("Failed to access global dry")
-                    .unwrap_or(false),
+                dry,
             )
             .with_context(|| {
                 format!("Failed to execute switch workspace with workspace_data {workspace_data:?}")
             })?;
         }
         Some(Active::Monitor(mid)) => {
-            switch_monitor(
-                mid,
-                global::OPTS
-                    .get()
-                    .map(|o| o.dry)
-                    .warn("Failed to access global dry")
-                    .unwrap_or(false),
-            )
-            .with_context(|| format!("Failed to execute switch monitor with monitor_id {mid:?}"))?;
+            switch_monitor(mid, dry).with_context(|| {
+                format!("Failed to execute switch monitor with monitor_id {mid:?}")
+            })?;
         }
         None => {
             warn!("Not executing switch (active = Unknown)");
